+++ src/EntityGraphQL/Schema/GqlTypeEnum.cs
@@ -10,4 +10,12 @@
         Mutation,
         Union
     }
+
+    public static class GqlTypesExtensions
+    {
+        public static bool IsNotValidForInput(this GqlTypes type)
+        {
+            return type == GqlTypes.Interface || type == GqlTypes.Mutation || type == GqlTypes.QueryObject || type == GqlTypes.Union;
+        }
+    }
 }