+++ src/tests/EntityGraphQL.Tests/QueryTests/InputTypeTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using EntityGraphQL.Schema;
 using Xunit;
 
@@ -111,6 +112,80 @@ public class InputTypeTests
         Assert.Contains((IEnumerable<dynamic>)((dynamic)result.Data["__type"]).fields, f => f.name == "unit");
         Assert.Contains((IEnumerable<dynamic>)((dynamic)result.Data["__type"]).fields, f => f.name == "dayOfWeek");
         Assert.Contains((IEnumerable<dynamic>)((dynamic)result.Data["__type"]).fields, f => f.name == "name");
+    }
+
+    class TaskInput : Task { }
+    class UserInput : User { }
+
+    [Fact]
+    public void SupportsQueryTypeReturnCorrectlyFindsTheInputTypeReturn()
+    {
+        var schema = SchemaBuilder.Create<TestDataContext>();
+
+        schema.AddType<User>("User");
+        schema.AddType<Task>("Task");
+
+        schema.UpdateType<User>(type =>
+        {
+            type.AddField("id", x => x.Id, null);
+            type.AddField("tasks", x => x.Tasks, null);
+        });
+
+        schema.UpdateType<Task>(type =>
+        {
+            type.AddField("id", x => x.Id, null);
+            type.AddField("user", null).Resolve<TestDataContext>((p, ctx) => ctx.Users.FirstOrDefault(u => u.Id == p.Project.CreatedBy));
+        });
+
+        var taskInput = schema.AddInputType<TaskInput>("TaskInput");
+        taskInput.AddField("id", x => x.Id, null);
+
+        var userInput = schema.AddInputType<UserInput>("UserInput");
+        userInput.AddField("id", x => x.Id, null);
+        // This returns Task but we should figure out we need TaskInput
+        userInput.AddField("tasks", x => x.Tasks, null);
+
+        schema.Query().AddField("users", (ctx) => ctx.Users, "Get a users");
+        schema.Query().AddField("tasks", (ctx) => ctx.Tasks, "Get a tasks");
+
+        var result = schema.ToGraphQLSchemaString();
+        Assert.Contains("input UserInput {\n\tid: Int!\n\ttasks: [TaskInput!]\n}", result);
+    }
+
+    [Fact]
+    public void SupportsQueryTypeAsInputTypeUsingReturns()
+    {
+        var schema = SchemaBuilder.Create<TestDataContext>();
+
+        schema.AddType<User>("User");
+        schema.AddType<Task>("Task");
+
+        schema.UpdateType<User>(type =>
+        {
+            type.AddField("id", x => x.Id, null);
+            type.AddField("tasks", x => x.Tasks, null);
+        });
+
+        schema.UpdateType<Task>(type =>
+        {
+            type.AddField("id", x => x.Id, null);
+            type.AddField("user", null).Resolve<TestDataContext>((p, ctx) => ctx.Users.FirstOrDefault(u => u.Id == p.Project.CreatedBy));
+        });
+
+        var taskInput = schema.AddInputType<Task>("TaskInput");
+        taskInput.AddField("id", x => x.Id, null);
+
+        var userInput = schema.AddInputType<User>("UserInput");
+        userInput.AddField("id", x => x.Id, null);
+        userInput.AddField("tasks", x => x.Tasks, null)
+            // Here we tell it what it returns
+            .Returns("TaskInput");
+
+        schema.Query().AddField("users", (ctx) => ctx.Users, "Get a users");
+        schema.Query().AddField("tasks", (ctx) => ctx.Tasks, "Get a tasks");
+
+        var result = schema.ToGraphQLSchemaString();
+        Assert.Contains("input UserInput {\n\tid: Int!\n\ttasks: [TaskInput!]\n}", result);
     }
 
     [GraphQLArguments]